+++ app/src/main/kotlin/ac/mdiq/podcini/playback/base/MediaPlayerBase.kt
@@ -53,6 +53,7 @@ import ac.mdiq.podcini.utils.Logpe
 import ac.mdiq.podcini.utils.Logps
 import ac.mdiq.podcini.utils.Logpt
 import ac.mdiq.podcini.utils.Logt
+import android.os.Build
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
@@ -955,6 +956,8 @@ abstract class MediaPlayerBase {
 
     companion object {
         private val TAG: String = MediaPlayerBase::class.simpleName ?: "Anonymous"
+
+        var enableFloat = Build.VERSION.SDK_INT >= 29
 
         private const val MIN_POSITION_SAVER_INTERVAL: Int = 5000   // in millisoconds